Replace requirejs with an easier to understand alternative
Just mocking the API to test this correctly. Note that I added "ok"
logging. It would be a good idea to capture the output and assert
against that. This is far better than the earlier solution, though.

The problem is that we still cannot prove that something failed to
execute. That would take different kind of control flow. Tests would
have to be responsible of shutting down Phantom and I'm not sure if I
like that idea.